+++ src/__init__.py
@@ -28,6 +28,7 @@ from src.distributions import (
 )
 from src.excel_export import gsm_allocation_to_dict, write_analysis_workbook
 from src.export import write_policy_comparison, write_summary_csv
+from src.eoq import EOQResult, compute_eoq, total_cost
 from src.fill_rate import (
     FillRateResult,
     fill_rate_from_inventory,
@@ -55,6 +56,7 @@ from src.newsvendor import (
     optimal_newsvendor_continuous_normal,
     optimal_newsvendor_discrete,
 )
+from src.powerbi_export import PowerBIDatasetPaths, build_powerbi_dataset
 from src.policies import PolicyResult, continuous_review_sq, periodic_review_rs
 from src.risk_period import RiskPeriodStats, demand_over_risk_period
 from src.safety_stock import SafetyStockResult, safety_stock, service_level_factor
@@ -80,6 +82,7 @@ __all__ = [
     "GammaParams",
     "NewsvendorResult",
     "PolicyResult",
+    "PowerBIDatasetPaths",
     "RSOptimizationResult",
     "RiskPeriodStats",
     "SQOptimizationResult",
@@ -87,6 +90,7 @@ __all__ = [
     "SimulationCostResult",
     "SimulationResult",
     "compare_review_periods",
+    "build_powerbi_dataset",
     "compute_eoq",
     "continuous_review_sq",
     "critical_ratio",